[MLIR][GPU] Add a dump-asm pass option to GPU serialization.
The dump-asm pass option allows developers to get the output of LLVM
for their kernels. This is usually the assembly of their kernel, hence
the name of the pass option. However, if they used an option such as
--stop-after, to extract intermediate IR from the partially-completed
compilation process. This simplifies the debugging of compilation
issues arising within the serialization of GPU code to binary blobs.

(Rationale: needing to go in and add a call to print the ASM out
and then take it back out again before commiting was getting to be a pain.)